Added a fragment  about group management module
The comparison of the taxonomic profiles between different groups of samples is a need to extract useful conclusions in many metagenomics studies. Tasks related with these analysis of groups, as the extraction of the minimal tree with all the taxa with some direct or accumulated assignment, will be defined and developed in the new in progress module.