Fragment for Biographika.
@edu please refine it and add ref
New visualization tools for metagenomics results are undoubtedly needed. Interactivity is a especially interesting feature visualizing metagenomics data since the expert needs to explore the results in a knowledge-driven way. The majority of the available metagenomics data visualizations are static. We are working in the project Biographika **cite Biorxive paper**, to provide interactive rich visualizations on the web for Bio4j data. The development of visualizations specific for MG7 is one of the Biographika current goals. Biographika uses D3.js, the de facto standard visualization JavaScript library, and is open source.